Start Qt Project
The Qt project was initialized using Qt Creator. The source, include and resources folders were created.
The MainWindow class, which is the application's Main Window, was initialized and a help menu with an "About" action and a registration menu with a "Client" action were inserted. In addition, some icons were inserted into the project.